requirements.txt
einops>=0.6.1
numpy>=1.24.4
open-clip-torch>=2.20.0
opencv-python==4.6.0.66
pillow>=9.5.0
scipy
timm
tokenizers

torch==2.1.2
torchmetrics==1.6.2
torchvision==0.16.2

tqdm>=4.65.0
transformers==4.46.3
triton==2.1.0
urllib3<1.27,>=1.25.4
xformers>=0.0.20
streamlit-keyup==0.2.0
lpips
peft
pyiqa
omegaconf
dominate
diffusers==0.25.1
gradio==3.43.1
huggingface_hub==0.25.0
neuralcompression